requirements.txt
flask 
requests
PyYAML
python-dotenv
PyJWT
cryptography